requirements.txt
# requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
boto3==1.34.0
python-multipart==0.0.6
pydantic==2.5.0
openai>=1.30.0
numpy>=1.24.0
PyPDF2>=3.0.0
python-docx>=0.8.11
python-pptx>=1.0.0
Pillow>=9.0.0
pytesseract>=0.3.10
pdf2image>=1.16.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt<4.0.0
httpx>=0.27.0
aiofiles>=23.0.0
requests>=2.31.0
python-magic-bin>=0.4.14
email-validator>=2.0.0